requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.4b2
bob.extension==7.0.3
bob.io.base==5.0.2
bob.learn.em==3.2.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
clip @ file:///home/niewei/pythonWork/FER/CLIP-main
cloudpickle==2.2.0
colorama==0.4.6
cycler==0.10.0
Cython==0.29.34
dask==2022.2.0
dask-glm==0.2.0
dask-ml==2022.5.27
decorator==5.1.1
detectron2 @ file:///home/niewei/Downloads/detectron2-0.6%2Bcu102-cp37-cp37m-linux_x86_64.whl
dill==0.3.6
distributed==2022.2.0
docker-pycreds==0.4.0
easydict==1.9
einops==0.4.1
filelock==3.12.0
fire==0.5.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20220512
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grad-cam==1.4.8
graphviz==0.8.4
grpcio==1.48.1
h5py==3.7.0
HeapDict==1.0.1
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
imageio==2.22.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
ipdb==0.13.13
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
lightning-utilities==0.10.1
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mdurl==0.1.2
menpo==0.11.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mmcv-full==1.4.3
mmdet==2.20.0
model-index==0.1.11
msgpack==1.0.4
multidict==6.0.5
multipledispatch==0.6.0
mxnet==1.8.0.post0
mypy-extensions==1.0.0
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.5.2.52
opencv-python-headless==4.9.0.80
openmim==0.3.7
ordered-set==4.1.0
overrides==7.3.1
packaging==21.3
pandas==1.3.5
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
parso==0.8.3
partd==1.3.0
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
portalocker==2.5.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydot==1.4.2
Pygments==2.15.1
pynndescent==0.5.12
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.9.5
pytorch-pretrained-vit==0.0.7
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.1.0
qudida==0.0.4
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.6.3
seaborn==0.12.2
sentry-sdk==1.23.1
setproctitle==1.3.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
supervision==0.6.0
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.0.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.12
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.10.0
torch-cluster @ file:///home/niewei/Downloads/torch_cluster-1.5.9-cp37-cp37m-linux_x86_64.whl
torch-geometric==2.1.0.post1
torch-scatter @ file:///home/niewei/Downloads/torch_scatter-2.0.9-cp37-cp37m-linux_x86_64.whl
torch-sparse @ file:///home/niewei/Downloads/torch_sparse-0.6.12-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ file:///home/niewei/Downloads/torch_spline_conv-1.2.1-cp37-cp37m-linux_x86_64.whl
torchaudio==0.10.0
torchfile==0.1.0
torchmetrics==0.11.4
torchnet==0.0.4
torchvision==0.11.0
tornado==6.1
tqdm==4.64.1
traitlets==5.3.0
transformers==4.29.2
ttach==0.0.3
typed-ast==1.5.4
typing_extensions==4.6.0
umap-learn==0.5.6
urllib3==2.0.2
visdom==0.1.8.9
wandb==0.15.3
warmup-scheduler==0.3
wcwidth==0.2.5
websocket-client==1.0.1
Werkzeug==2.2.2
xmltodict==0.13.0
yacs==0.1.8
yapf==0.33.0
yarl==1.9.4
zict==2.2.0
zipp==3.15.0
